requirements.txt
pymongo
pinecone-client
scikit-learn
matplotlib
numpy
seaborn
plotly
pandas
dash